Optimism: uniformising the imported/defined types
It seems we do redefine at multiple place the same types/structures.
Uniformising it in this commit
